requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
aiosqlite==0.20.0
